Fix several issues in the migrator
